Realizamos algunos cambios por erorres

+++ src/main/java/org/svalero/facilreserva/service/ReservationService.java
@@ -1,5 +1,6 @@
 package org.svalero.facilreserva.service;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.svalero.facilreserva.domain.Reservation;
@@ -12,23 +13,34 @@ import java.util.List;
 public class ReservationService {
     @Autowired
     private ReservationRepository reservationRepository;
+
     @Autowired
-    private modelMapper modelMapper;
+    private ModelMapper modelMapper;
 
     public List<Reservation> getAll() {
         return reservationRepository.findAll();
     }
+
     public Reservation get(long reservationId) throws ReservationNotFoundException {
-        return reservationRepository.findById(reservationId).orElseThrow(ReservationNotFoundException::new);
+        return reservationRepository.findById(reservationId)
+                .orElseThrow(ReservationNotFoundException::new);
     }
 
-    public Reservation modify(Reservation reservation) {
-        
+    public Reservation modify(long reservationId, Reservation reservation) throws ReservationNotFoundException {
+        Reservation existingReservation = reservationRepository.findById(reservationId)
+                .orElseThrow(ReservationNotFoundException::new);
+
+        modelMapper.map(reservation, existingReservation);
+        return reservationRepository.save(existingReservation);
     }
+
     public void delete(long reservationId) throws ReservationNotFoundException {
-        reservationRepository.deleteById(reservationId).orElseThrow(ReservationNotFoundException::new);
+        if (!reservationRepository.existsById(reservationId)) {
+            throw new ReservationNotFoundException();
+        }
         reservationRepository.deleteById(reservationId);
     }
+
     public Reservation add(Reservation reservation) {
         return reservationRepository.save(reservation);
     }
